% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasm_methods.R
\name{autoplot.seasm}
\alias{autoplot.seasm}
\title{Autoplot method for seasonality GAM on monthly counts}
\usage{
\method{autoplot}{seasm}(object, ..., type = "fit")
}
\arguments{
\item{object}{an object of class "seasm".}

\item{...}{other plotting parameters (not used in this function)}

\item{type}{a character denoting what type of plot should be drawn. Defaults to "rating_curve". Possible types are
\itemize{
   \item{"fit"}{ to plot the GAM fit to the monthly counts}
   \item{"seasonality"}{ to plot the seasonal smooth term in th GAM model}
   \item{"annual"}{ to plot the annual smooth term in th GAM model}
}}
}
\value{
returns an object of class "ggplot2".
}
\description{
Visualize different aspects of the seasonality GAM
}
\seealso{
\code{\link{seasonality_gam}} and \code{\link{summary.seasm}}, \code{\link{plot.seasm}}
}
