% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diagnosis_data}
\alias{diagnosis_data}
\title{Simulated diagnosis data}
\format{
A data frame with columns:
\describe{
 \item{ID}{ID of the individual}
 \item{ICD}{ICD code of the disease endpoint}
 \item{EVENT_DATE}{Date of diagnosis}
}
}
\usage{
diagnosis_data
}
\description{
Diagnosis data simulated using a sinusoidal curve
}
\keyword{datasets}
