% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_phenotype.R
\name{get_seasonality_phenotype}
\alias{get_seasonality_phenotype}
\title{Calculate seasonality phenotype based on seasonality GAM}
\usage{
get_seasonality_phenotype(
  mod,
  data = NULL,
  year_start = NULL,
  year_end = NULL,
  adjusted = F
)
}
\arguments{
\item{mod}{object of type "seasm"}

\item{data}{Optional data.frame with columns:
      \itemize{
         \item{"ID"}{ a character denoting the ID of the patient}
         \item{"EVENT_DATE"}{ character in the format YYYY-MM-DD denoting the first date the disease is recorded in the registry}
      }
Useful if mod was fitted on an external data set.}

\item{year_start}{integer denoting the starting year of the analysis}

\item{year_end}{integer denoting the final year of the analysis}

\item{adjusted}{boolean denoting whether to perform specific adjustments for months with low hospital usage (July and December). Defaults to FALSE.}

\item{...}{Not used for this function}
}
\value{
The function returns a tibble containing two seasonality phenotype:
                   \itemize{
                      \item{"seasonal_val_binary"}{ Dichotomous value indicating whether an individual was diagnosed during high/low season as determined by mod}
                      \item{"seasonal_val_01"}{ The proportional height of the seasonality pattern from mod at the diagnosis date}
                   }
}
\description{
Based on an input model, this function calculates the seasonality phenotype for each individual
}
\details{
Details here
}
\seealso{
\code{\link{seasonality_gam}}
}
