% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD_codes}
\alias{ICD_codes}
\title{ICD codes for seasonality replication study}
\format{
A data frame with columns:
\describe{
 \item{disease_endpoint}{ID of the disease endpoint}
 \item{ICD_version}{Version of ICD}
 \item{ICD}{ICD code}
}
}
\usage{
ICD_codes
}
\description{
ICD code table with codes used for the seasonality replication study
}
\keyword{datasets}
