% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_modeling.R
\name{prepare_output}
\alias{prepare_output}
\title{Prepare output for a seasm model object}
\usage{
prepare_output(x, path = "output.RData")
}
\arguments{
\item{x}{a named list of seasonality model objects, where the names represent the corresponding the diseases}

\item{path}{path to the RData file to be written to disk
\itemize{
   \item{'-'}{ the data field has been removed, to ensure no individual-level data is included}
   \item{"-"}{ the monthly_counts table has been curated such that only months where 5 or more individuals are diagnosed are included.}
   \item{"-"}{ the monthly counts table contained in both fit and fit_null has been removed}
}}

\item{...}{Not used for this function}
}
\value{
No return value, called for side effects. Saves an RData file containing a curated version of the seasm model objects where
}
\description{
Package the seasonality model to a file to send back the results
}
\seealso{
\code{\link{seasonality_gam}}, \code{\link{summary.seasm}}, \code{\link{plot.seasm}}
}
