% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasm_methods.R
\name{print.seasm}
\alias{print.seasm}
\title{Print method for seasonality models}
\usage{
\method{print}{seasm}(x, ...)
}
\arguments{
\item{x}{an object of class "seasm".}

\item{...}{Not used for this function}
}
\description{
Print the summary of the seasonality model
}
\seealso{
\code{\link{seasonality_gam}}
}
