% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_modeling.R
\name{seasonality_gam}
\alias{seasonality_gam}
\title{Run seasonality model using monthly counts data}
\usage{
seasonality_gam(
  data = NULL,
  monthly_counts = NULL,
  year_start = NULL,
  year_end = NULL,
  adjusted = F
)
}
\arguments{
\item{data}{data.frame with the diagnosis date for each individual with the disease. It should contain the columns:
\itemize{
   \item{"ID"}{ a character denoting the ID of the patient}
   \item{"EVENT_DATE"}{ character in the format YYYY-MM-DD denoting the first date the disease is recorded in the registry}
}}

\item{monthly_counts}{data.frame with the diagnosis counts per year per month. It should contain the columns:
\itemize{
   \item{"EVENT_YEAR"}{ Integer denoting the year}
   \item{"EVENT_DATE"}{ Integer denoting the month}
   \item{"COUNT"}{ Integer denoting the number of individuals diagnosed in a particular year and month}
}
Either data or monthly_counts must be provided. If both are provided, monthly_counts argument will be ignored and data used to derive the monthly_counts.}

\item{year_start}{integer denoting the starting year of the analysis.
Defaults to NULL in which case the starting year is taken to be the earliest year in data}

\item{year_end}{integer denoting the final year of the analysis.
Defaults to NULL in which case the final year is taken to be the latest year in data}

\item{adjusted}{boolean denoting whether to perform specific adjustments for months with low hospital usage (July and December). Defaults to FALSE.}

\item{...}{Not used for this function}
}
\value{
The function returns an object of type "seasm", including the objects:
 \item{\code{fit}}{ GAM seasonality model fit object which was inferred using aggregated monthly counts data}
 \item{\code{fit_null}}{GAM null model fit object, which does not include a seasonality smooth term, also inferred using aggregated monthly counts data}
 \item{\code{fit_table}}{Table with the fitted values for the seasonality and the null model}
 \item{\code{seasonality_term}}{Table with the values of the seasonality smooth term from the seasonality model}
 \item{\code{annual_term}}{Table with the values of the annual smooth term from the seasonality model}
 \item{\code{summary}}{Summary statistics of the seasonality curve}
 \item{\code{monthly_counts}}{Diagnosis counts per month, derived from the input data.frame}
 \item{\code{data}}{The input data.frame}
 \item{\code{year_start}}{input argument or derived from data}
 \item{\code{year_end}}{input argument or derived from data}
 \item{\code{adjusted}}{input argument}
}
\description{
Runs a GAM which includes an annual and a seasonal term assuming a quasipoisson count distribution
}
\details{
Details here
}
\seealso{
\code{\link{summary.seasm} and \link{plot.seasm}} for further studying the output
}
