% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasm_methods.R
\name{summary.seasm}
\alias{summary.seasm}
\title{Summary method for seasonality models}
\usage{
\method{summary}{seasm}(object, ...)
}
\arguments{
\item{object}{an object of class "seasm".}

\item{...}{Not used for this function}
}
\description{
Summarize a seasonality gam model object
}
\seealso{
\code{\link{seasonality_gam}},
}
