% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_phenotype.R
\name{get_seasonality_phenotype}
\alias{get_seasonality_phenotype}
\title{Calculate seasonality phenotype based on seasonality GAM}
\usage{
get_seasonality_phenotype(mod, year_start, year_end)
}
\arguments{
\item{mod}{object of type "seasm"}

\item{year_start}{integer denoting the starting year of the analysis}

\item{year_end}{integer denoting the final year of the analysis}

\item{...}{Not used for this function}
}
\value{
The function returns a tibble containing two seasonality phenotype:
                   \itemize{
                      \item{"seasonal_val_binary"}{ Dichotomous value indicating whether an individual was diagnosed during high/low season as determined by mod}
                      \item{"seasonal_val_01"}{ The proportional height of the seasonality pattern from mod at the diagnosis date}
                   }
}
\description{
Based on an input model, this function calculates the seasonality phenotype for each individual
}
\details{
Details here
}
\seealso{
\code{\link{seasonality_gam}}
}
