% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_phenotype.R
\name{get_seasonality_phenotype}
\alias{get_seasonality_phenotype}
\title{Calculate seasonality phenotype based on seasonality GAM}
\usage{
get_seasonality_phenotype(mod, data = NULL, year_start = NULL, year_end = NULL)
}
\arguments{
\item{mod}{object of type "seasm"}

\item{data}{Optional data.frame with columns:
      \itemize{
         \item{"ID"}{ a character denoting the ID of the patient}
         \item{"EVENT_DATE"}{ character in the format YYYY-MM-DD denoting the first date the disease is recorded in the registry}
      }
Useful if mod was fitted on an external data set.}

\item{year_start}{integer denoting the starting year of the analysis}

\item{year_end}{integer denoting the final year of the analysis}

\item{...}{Not used for this function}
}
\value{
The function returns a tibble containing two seasonality phenotype:
                   \itemize{
                   \item{"ID"}{ unique sample ID in your biobank}
                   \item{"EVENT_DATE"}{ first recorded date of diagnosis of the disease in the format YYYY-MM-DD}
                   \item{"EVENT_MONTH_DEC"}{ float denoting the event date as decimal number of months}
                   \item{"pheno_binary"}{ dichotomous seasonality phenotype value. 1 denotes that the individual was diagnosed in the season of high incidence and 0 denotes that the individual was diagnosed in the season of low incidence}
                   \item{"pheno_qt"}{ quantitative seasonality phenotype value. This denotes the extent of the seasonality curve when an individual is diagnosed with a disease. The phenotype has been quantile normalized.}
                   }
}
\description{
Based on an input seasonality model, this function calculates the seasonality phenotype for each individual
}
\details{
Details here
}
\seealso{
\code{\link{seasonality_gam}}
}
