% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_modeling.R
\name{seasonality_gam}
\alias{seasonality_gam}
\title{Run seasonality model using monthly counts data}
\usage{
seasonality_gam(data, year_start = NULL, year_end = NULL, adjusted = F)
}
\arguments{
\item{data}{data.frame with columns:
\itemize{
   \item{"ID"}{ a character denoting the ID of the patient}
   \item{"EVENT_DATE"}{ The first date when the ICD code is found in the registry}
}}

\item{year_start}{integer denoting the starting year of the analysis.
Defaults to NULL in which case the starting year is taken to be the earliest year in data}

\item{year_end}{integer denoting the final year of the analysis.
Defaults to NULL in which case the final year is taken to be the latest year in data}

\item{adjusted}{boolean denoting whether to perform seasonal behavioural adjustment. Defaults to FALSE}

\item{...}{Not used for this function}
}
\value{
The function returns an object of type "seasm"
}
\description{
Runs a GAM which includes an annual and a seasonal term assuming a quasipoisson count distribution
}
\details{
Details here
}
\seealso{
\code{\link{summary.seasm} and \link{plot.seasm}} for further studying the output
}
