% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_modeling.R
\name{seasonality_gam}
\alias{seasonality_gam}
\title{Run seasonality model using monthly counts data}
\usage{
seasonality_gam(data, ICD_vec, year_start, year_end)
}
\arguments{
\item{data}{data.frame with columns:
\itemize{
   \item{"ID"}{ a character denoting the ID of the patient}
   \item{"ICD"}{ ICD code}
   \item{"EVENT_DATE"}{ The first date when the ICD code is found in the registry}
}}

\item{ICD_vec}{character vector with the ICD codes characterizing the disease to be analyzed}

\item{year_start}{integer denoting the starting year of the analysis}

\item{year_end}{integer denoting the final year of the analysis}

\item{...}{Not used for this function}
}
\value{
The function returns an object of type "seasm"
}
\description{
Runs a GAM which includes an annual and a seasonal term assuming a quasipoisson count distribution
}
\details{
Details here
}
\seealso{
\code{\link{summary.seasm} and \link{plot.seasm}} for further studying the output
}
